host/requirements.txt
Pillow==9.2.0
pyperclip==1.8.2